add go16 et go17 file due to req.withcontext which is only go 1.7
